requirements.txt
# CreatorStudio AI - Professional Content Creation Suite
# Requirements for Windows 11 Python Application

# Core GUI Framework
PyQt6==6.6.1
PyQt6-WebEngine==6.6.0

# AI and Machine Learning
openai==1.12.0
anthropic==0.18.1
transformers==4.37.2
torch==2.2.0
pillow==10.2.0

# Video Processing
moviepy==1.0.3
opencv-python==4.9.0.80
imageio==2.33.1
imageio-ffmpeg==0.4.9

# Image Processing
numpy==1.26.3
scikit-image==0.22.0
matplotlib==3.8.2
plotly==5.18.0

# Audio Processing
pyttsx3==2.90
pydub==0.25.1
SpeechRecognition==3.10.1

# Data and Database
SQLAlchemy==2.0.25
pandas==2.2.0
openpyxl==3.1.2

# Web and API
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3

# Social Media APIs
tweepy==4.14.0
instagrapi==2.0.0

# Analytics and Charts
seaborn==0.13.2
wordcloud==1.9.3

# Utilities
python-dotenv==1.0.1
colorama==0.4.6
tqdm==4.66.1
pyyaml==6.0.1
jsonschema==4.21.1
cryptography==42.0.2
schedule==1.2.1

# Development
black==24.1.1
pytest==8.0.0
